chore: remove calls to payment service
As exaplained in https://docs.google.com/document/d/1QUTdG1V4uHu5Ov8YIMRd9vJB62RmtsJodB5HJGXmftw/edit#, we have a bunch of billing related code that's incomplete or unused.

In particular the payment-service was only used by building blocks, however that was currently broken as it was launching a CORS error.

Furthermore, we don't really know how this interact with the legacy billing service we have, since it creates payment entries and configuration that's not replicated to the engine. meaning that it won't appear on the manager, and probably won't even considered at the moment of paying out to customers.