fix: schema issues.
note: back when the campaign models were created for the modal, fields like "bid_count" I believe didn't exist on the campaign service campaign model, therefore we included it manually on the campaign response to the modal. A few months ago this bid count was added to the campaign service model under a slightly different name (activeBidsCount vs bidCount). This might imply a refactor.